Move DeviceAuthenticator into Authentication
