+++ visitz/Visitz/Services/Authentication/DeviceAuthenticator.cs
@@ -2,7 +2,7 @@
 using Plugin.Fingerprint;
 using Plugin.Fingerprint.Abstractions;
 
-namespace Visitz.Services
+namespace Visitz.Services.Authentication
 {
     /// <summary>
     /// Authenticate a user via fingerprint, face id or any other biometic / local authentication method.
